// src/service/file/schemas/studentRecordCommentForAdd.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

/**
 * 관찰 평가 결과 영역별 내용들
 * @nullable
 */
export type StudentRecordCommentForAdd = {
  /**
   * 영역 구분 코드
   * @maxLength 3
   */
  evaluationDomainCode: string;
  /**
   * 평가지표 내용
   * @maxLength 500
   */
  evaluationDescription: string;
  /** 유저가 수정했는지 여부 */
  userEdited: boolean;
} | null;