// src/service/file/schemas/studentRecordCommentResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */

/**
 * 관찰 평가 영역별 내용들
 */
export interface StudentRecordCommentResult {
  /** 평가 코멘트에 대한 아이디 */
  id: number;
  /**
   * 관찰 기록의 id
   * @minimum 0
   */
  studentRecordId: number;
  /**
   * 관찰 기록 대상 학생의 id
   * @minimum 0
   */
  studentId: number;
  /** 영역 구분 */
  evaluationDomainCode: string;
  /**
   * 평가지표 내용
   * @maxLength 500
   */
  evaluationDescription: string;
  /** 유저가 수정했는지 여부 */
  userEdited: boolean;
  /** 생성일 */
  createdAt: string;
}